refactor: one-click skill export + sidebar panel — no modal
Fewer clicks, no dialog:

- The toolbar package icon (and Cmd+Shift+K) now downloads the .skill
  immediately, like the other export buttons — smart-defaulted, validated
  metadata; the toast carries the add-to-Claude steps with a copy-command
  action.
- Everything optional moved to an always-visible "Agent Skill" sidebar
  card: inline-editable name/description (validated as you type, one
  contextual hint), Instructions/Knowledge toggle, AI refine, and a
  collapsed import/template section (file, GitHub URL, marketplace
  quick-picks, discovery picker).
- New SkillMetaProvider shares the edited metadata between the toolbar
  export and the panel; ContentContext gains a document-replacer registry
  so the sidebar can import through the page's undo-friendly edit path.
- Removes the wizard dialog entirely.

Verified in-browser: panel pre-fills, renaming in the panel flows to the
one-click toolbar download (renamed-from-panel.skill). 60 tests, lint,
build clean.
