+++ src/contexts/content-context.tsx
@@ -1,18 +1,45 @@
 "use client";
 
-import { createContext, useContext, useMemo, useState, ReactNode } from "react";
+import {
+  createContext,
+  useCallback,
+  useContext,
+  useMemo,
+  useRef,
+  useState,
+  ReactNode,
+} from "react";
 
 interface ContentContextType {
   content: string;
   setContent: (content: string) => void;
+  /** Replace the editor document (no-op until the editor registers a handler). */
+  replaceDocument: (markdown: string) => void;
+  /** Called by the editor page to receive document replacements (import, template). */
+  registerDocumentReplacer: (handler: ((markdown: string) => void) | null) => void;
 }
 
 const ContentContext = createContext<ContentContextType | undefined>(undefined);
 
 export function ContentProvider({ children }: { children: ReactNode }) {
   const [content, setContent] = useState("");
+  const replacerRef = useRef<((markdown: string) => void) | null>(null);
 
-  const value = useMemo(() => ({ content, setContent }), [content]);
+  const replaceDocument = useCallback((markdown: string) => {
+    replacerRef.current?.(markdown);
+  }, []);
+
+  const registerDocumentReplacer = useCallback(
+    (handler: ((markdown: string) => void) | null) => {
+      replacerRef.current = handler;
+    },
+    [],
+  );
+
+  const value = useMemo(
+    () => ({ content, setContent, replaceDocument, registerDocumentReplacer }),
+    [content, replaceDocument, registerDocumentReplacer],
+  );
 
   return (
     <ContentContext.Provider value={value}>
